analytics user tab done
