+++ src/components/index.ts
@@ -41,3 +41,6 @@ export { default as Transform } from "./Transform";
 export { default as UserProSideBar } from "./UserProSideBar";
 export { default as Version } from "./Version";
 export { default as Widget } from "./Widget";
+export { default as AnalyticsTab } from "./AnalyticsTab";
+export { default as AnalyticsUser } from "./AnalyticsUser";
+
